refactor: Split template and styles from component class And change file names according to ng conventions
