+++ src/dropdown/dropdown.component.css
@@ -1,0 +1,7 @@
+a {
+  outline: none !important;
+}
+
+.dropdown-inline {
+  display: inline-block;
+}
