+++ src/dropdown/dropdown.module.ts
@@ -1,0 +1,12 @@
+import { MultiselectDropdown } from './dropdown.component';
+import { MultiSelectSearchFilter } from './search-filter.pipe';
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+import { FormsModule } from '@angular/forms';
+
+@NgModule({
+  imports: [CommonModule, FormsModule],
+  exports: [MultiselectDropdown, MultiSelectSearchFilter],
+  declarations: [MultiselectDropdown, MultiSelectSearchFilter],
+})
+export class MultiselectDropdownModule { }
