+++ src/dropdown/search-filter.pipe.ts
@@ -1,0 +1,18 @@
+import { Pipe, PipeTransform } from '@angular/core';
+import { IMultiSelectOption } from './types';
+
+@Pipe({
+  name: 'searchFilter'
+})
+export class MultiSelectSearchFilter implements PipeTransform {
+  transform(options: Array<IMultiSelectOption>, args: string): Array<IMultiSelectOption> {
+    const matchPredicate = (option: IMultiSelectOption) => option.name.toLowerCase().indexOf((args || '').toLowerCase()) > -1,
+      getChildren = (option: IMultiSelectOption) => options.filter(child => child.parentId === option.id),
+      getParent = (option: IMultiSelectOption) => options.find(parent => option.parentId === parent.id);
+    return options.filter((option: IMultiSelectOption) => {
+      return matchPredicate(option) ||
+        (typeof (option.parentId) === 'undefined' && getChildren(option).some(matchPredicate)) ||
+        (typeof (option.parentId) !== 'undefined' && matchPredicate(getParent(option)));
+    });
+  }
+}
