+++ src/dropdown/types.ts
@@ -1,0 +1,35 @@
+export interface IMultiSelectOption {
+  id: any;
+  name: string;
+  isLabel?: boolean;
+  parentId?: any;
+  params?: any;
+}
+
+export interface IMultiSelectSettings {
+  pullRight?: boolean;
+  enableSearch?: boolean;
+  checkedStyle?: 'checkboxes' | 'glyphicon' | 'fontawesome';
+  buttonClasses?: string;
+  itemClasses?: string;
+  containerClasses?: string;
+  selectionLimit?: number;
+  closeOnSelect?: boolean;
+  autoUnselect?: boolean;
+  showCheckAll?: boolean;
+  showUncheckAll?: boolean;
+  fixedTitle?: boolean;
+  dynamicTitleMaxItems?: number;
+  maxHeight?: string;
+  displayAllSelectedText?: boolean;
+}
+
+export interface IMultiSelectTexts {
+  checkAll?: string;
+  uncheckAll?: string;
+  checked?: string;
+  checkedPlural?: string;
+  searchPlaceholder?: string;
+  defaultTitle?: string;
+  allSelected?: string;
+}
