+++ src/index.ts
@@ -1,1 +1,4 @@
-export * from './multiselect-dropdown';
+export * from './dropdown/types';
+export * from './dropdown/search-filter.pipe';
+export * from './dropdown/dropdown.module';
+export * from './dropdown/dropdown.component';
